backend/auth-service/requirements.txt
# Core dependencies
Flask==2.3.3
flask-cors==4.0.0
flask-restful==0.3.10
pydantic==2.4.2
SQLAlchemy==2.0.22
psycopg2-binary==2.9.9
alembic==1.12.0

# Authentication
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==41.0.4

# Validation and security
email-validator==2.0.0
python-dotenv==1.0.0
marshmallow==3.20.1

# Production dependencies
gunicorn==21.2.0
prometheus-flask-exporter==0.22.4
sentry-sdk==1.32.0

# Testing
pytest==7.4.2